+++ gakumas-tools/simulator/CHANGELOG.md
@@ -8,6 +8,12 @@ Dates are based on Eastern Time Zone.
 
 
 
+
+## 2025-11-16
+
+- Initial implementation of link contest
+- Update Tsubame
+
 ## 2025-10-31
 
 - Update Mite Mao